requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scikit-learn>=1.3.0
Pillow>=10.0.0
opencv-python>=4.8.0

# XAI
grad-cam>=1.4.8          # pytorch-grad-cam
lime>=0.2.0.1
shap>=0.43.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Utilities
pyyaml>=6.0.1
tqdm>=4.66.0
matplotlib>=3.7.0
seaborn>=0.13.0
pandas>=2.1.0
scipy>=1.11.0

# Evaluation / Metrics
torchmetrics>=1.2.0

# Logging & Experiment Tracking
tensorboard>=2.14.0

# Dev
pytest>=7.4.0
black>=23.9.0
